Refactor fairness eval into modules
